% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{ANOVA}
\alias{ANOVA}
\title{Wrapper for summary(aov)}
\usage{
ANOVA(formula)
}
\arguments{
\item{formula}{An anova formula (variable ~ grouping variable)}
}
\description{
Wrapper for summary(aov)
}
\seealso{
stats::aov
}
