% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{as.data.frame.desctable}
\alias{as.data.frame.desctable}
\title{As.data.frame method for desctable}
\usage{
\method{as.data.frame}{desctable}(x, ...)
}
\arguments{
\item{x}{A desctable}

\item{...}{Additional as.data.frame parameters}
}
\value{
A flat dataframe
}
\description{
As.data.frame method for desctable
}
\keyword{deprecated}
