% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{datatable.desctable}
\alias{datatable.desctable}
\title{datatable method for desctable}
\usage{
\method{datatable}{desctable}(data = NULL, ...)
}
\arguments{
\item{data}{A desctable}

\item{...}{Additional datatable parameters}
}
\description{
datatable method for desctable
}
