% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{flatten_desctable}
\alias{flatten_desctable}
\title{Flatten a desctable to a dataframe recursively}
\usage{
flatten_desctable(desctable)
}
\arguments{
\item{desctable}{A desctable object}
}
\value{
A flat dataframe
}
\description{
Flatten a desctable to a dataframe recursively
}
\keyword{deprecated}
\keyword{internal}
