% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_quote}
\alias{formula_quote}
\title{Quote an argument in a formula if it is not already a formula}
\usage{
formula_quote(deparsed, orig)
}
\arguments{
\item{deparsed}{The unevaluated argument (with rlang::enexprs)}

\item{orig}{The evaluated argument}
}
\value{
A formula
}
\description{
The argument must have been captured unevaluated first
}
