% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{head_dataframe}
\alias{head_dataframe}
\title{Build the header for dataframe}
\usage{
head_dataframe(head)
}
\arguments{
\item{head}{A headerList object}
}
\value{
A names vector
}
\description{
Build the header for dataframe
}
\keyword{deprecated}
\keyword{internal}
