% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{head_datatable}
\alias{head_datatable}
\title{Build the header for datatable}
\usage{
head_datatable(head)
}
\arguments{
\item{head}{A headerList object}
}
\value{
An htmltools$tags object containing the header
}
\description{
Build the header for datatable
}
\keyword{deprecated}
\keyword{internal}
