% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{head_pander}
\alias{head_pander}
\title{Build the header for pander}
\usage{
head_pander(head)
}
\arguments{
\item{head}{A headerList object}
}
\value{
A names vector
}
\description{
Build the header for pander
}
\keyword{deprecated}
\keyword{internal}
