% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{header}
\alias{header}
\title{Build header}
\usage{
header(desctable, output = c("pander", "datatable", "dataframe"))
}
\arguments{
\item{desctable}{A desctable object}

\item{output}{An output format for the header}
}
\value{
A header object in the output format
}
\description{
Take a desctable object and create a suitable header for the mentionned output.
Output can be one of "pander", "datatable", or "dataframe".
}
\keyword{deprecated}
\keyword{internal}
