% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headerList}
\alias{headerList}
\title{Build a header list object}
\usage{
headerList(desctable)
}
\arguments{
\item{desctable}{A desctable}
}
\value{
A nested list of headers with colspans
}
\description{
Build a header list object
}
