% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert}
\alias{insert}
\title{Insert a vector y inside another vector x at position}
\usage{
insert(x, y, position)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector or list of vectors}

\item{position}{The position / vector of positions to insert vector(s) y in vector x}
}
\value{
The combined vector
}
\description{
Insert a vector y inside another vector x at position
}

