% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{IQR}
\alias{IQR}
\title{Return the inter-quartile range}
\usage{
IQR(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The inter-quartile range of the vector
}
\description{
Return the inter-quartile range of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

