% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_normal}
\alias{list_normal}
\title{List the normal variables in the dataframe}
\usage{
list_normal(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A list of variable names
}
\description{
List the normal variables in the dataframe
}
