% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_param}
\alias{list_param}
\title{List the parametric variables in the dataframe}
\usage{
list_param(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A list of variable names
}
\description{
List the parametric variables in the dataframe
}

