% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Max}
\alias{Max}
\title{Return the maximal value}
\usage{
Max(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The maximal value of the vector
}
\description{
Return the maximal value of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

