% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Mean}
\alias{Mean}
\title{Return the mean}
\usage{
Mean(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The mean of the vector
}
\description{
Return the mean of the numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

