% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Med}
\alias{Med}
\title{Return the median}
\usage{
Med(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The median of the vector
}
\description{
Return the median of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

