% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Min}
\alias{Min}
\title{Return the minimal value}
\usage{
Min(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The minimal value of the vector
}
\description{
Return the minimal value of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

