% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{N_}
\alias{N_}
\title{Return the number of observations}
\usage{
N_(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of sample sizes
}
\description{
Return the number of observations.
For numerical values, return the sample size. Missing values are omitted.
For factors, return the sample size for each level.
}

