% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{N_mean_}
\alias{N_mean_}
\title{Return the number of observations / mean}
\usage{
N_mean_(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of sample sizes
}
\description{
Return the mean of a numeric vector.
For numerical values, return the sample size
NAs are removed by default.
For factors, return the sample size for each level
}

