% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pander.desctable}
\alias{pander.desctable}
\title{Pander method for desctable}
\usage{
pander.desctable(x = NULL, ...)
}
\arguments{
\item{x}{A desctable}

\item{...}{Additional pander parameters}
}
\description{
Pander method for desctable
}
