% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse a formula}
\usage{
parse_formula(x, f)
}
\arguments{
\item{x}{The variable to test it on}

\item{f}{A formula to parse}
}
\value{
A function to use as a stat/test
}
\description{
Parse a formula defining the conditions to pick a stat/test
}
\details{
Parse a formula defining the conditions to pick a stat/test
and return the function to use.
The formula is to be given in the form of
conditional ~ T | F
and conditions can be nested such as
conditional1 ~ (conditional2 ~ T | F) | F
The FALSE option can be omitted, and the TRUE can be replaced with NA
}
\examples{
# To use one column for different statistics depending on the variable
parse_formula(is.factor ~ percent | (is.normal ~ mean | median), factor(rep(LETTERS[1:3], 5)))

# To use one column selectively for a type of variable
parse_formula(is.normal ~ mean, rnorm(1000))
}
