% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Pct}
\alias{Pct}
\title{Return the percentage of observations}
\usage{
Pct(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of percentages
}
\description{
Return the percentage of observations.
For numerical values, return NA.
For factors, return the percentage of each level.
}

