% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{percent}
\alias{percent}
\title{Return the percentages for the levels of a factor}
\usage{
percent(x)
}
\arguments{
\item{x}{A factor}
}
\value{
A nlevels(x) + 1 length vector of percentages
}
\description{
Return a compatible vector of length nlevels(x) + 1
to print the percentages of each level of a factor
}
