% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{petrify}
\alias{petrify}
\title{Petrifies a table for output}
\usage{
petrify(data, digits = 2, ...)
}
\arguments{
\item{data}{The structure returned by desctable}

\item{digits}{Argument for prettyNum. Defaults to 2}

\item{...}{Arguments to pass to prettyNum}
}
\value{
A petrified desctable
}
\description{
Using prettyNum, it turns a desctable to stone as a character dataframe
}
\details{
This function turns a desctable into stone as a single character dataframe.
It combines the multiple tables for a group'ed desctable into one, with headers.
The ... arguments are passed to prettyNum to pretty-print the numeric columns.
The 'digits' argument is passed with a value of 2 by default, but can be overriden.

The result being a less mutable dataframe (thus turned to stone), with prettified numbers,
it is 'petrified'.
}
\seealso{
\code{\link{desctable}}
}
