% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{print.desctable}
\alias{print.desctable}
\title{Print method for desctable}
\usage{
\method{print}{desctable}(x, ...)
}
\arguments{
\item{x}{A desctable}

\item{...}{Additional print parameters}
}
\value{
A flat dataframe
}
\description{
Print method for desctable
}
