% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Q1}
\alias{Q1}
\title{Return the first quartile}
\usage{
Q1(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The first quartile of the vector
}
\description{
Return the first quartile of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

