% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Q3}
\alias{Q3}
\title{Return the third quartile}
\usage{
Q3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The third quartile of the vector
}
\description{
Return the third quartile of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

