% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Range}
\alias{Range}
\title{Return the range}
\usage{
Range(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The range of the vector
}
\description{
Return the range of a numeric vector.
NAs are removed by default.
For non-numerical vectors, return NA.
}

