% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Sd}
\alias{Sd}
\title{Return the standard deviation}
\usage{
Sd(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The standard deviation of the vector
}
\description{
Return the standard deviation of a numeric vectorr
NAs are removed by default.
For non-numerical vectors, return NA.
}

