% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{set_desctable_class}
\alias{set_desctable_class}
\title{Set the "desctable" class to the passed object}
\usage{
set_desctable_class(x)
}
\arguments{
\item{x}{Object to set the "desctable" class to}
}
\value{
The object with the class "desctable"
}
\description{
Set the "desctable" class to the passed object
}
\keyword{deprecated}
\keyword{internal}
