% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{simpleTable}
\alias{simpleTable}
\title{Generate a simple statistics table}
\usage{
simpleTable(data, stats, labels = NULL)
}
\arguments{
\item{data}{The dataframe to analyse}

\item{stats}{A list of statistics to apply to each element of the dataframe}

\item{labels}{A named character vector of labels to use instead of variable names}
}
\value{
A table of statistics for all variables
}
\description{
Generate a simple statistics table with variable names/labels and levels
}
\details{
labels is an option named character vector used to make the table prettier.
If given, the variable names for which there is a label will be replaced by their corresponding label.
Not all variables need to have a label, and labels for non-existing variables are ignored.

The resulting dataframe is directly pipe-able to pander or DT, or can be exported like any dataframe to csv, etc.
}

