% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{statColumn}
\alias{statColumn}
\title{Generate one statistic for all variables}
\usage{
statColumn(stat, data)
}
\arguments{
\item{stat}{The statistic to use}

\item{data}{The dataframe to apply the statistic to}
}
\value{
A vector for one statistic column
}
\description{
Use one stat function (made safe using statify) on all the data
to produce a single statistics column.
}
\details{
The result is either a numeric vector, or a character vector if
the content of the column is not made entirely of numbers.
}
\keyword{internal}
