% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statify}
\alias{statify}
\title{Transform any function into a valid stat function for the table}
\usage{
statify(x, f)
}
\arguments{
\item{x}{A vector}

\item{f}{The function to try to apply, or a formula combining two functions}
}
\value{
The results for the function applied on the vector, compatible with the format of the result table
}
\description{
Transform a function into a valid stat function for the table
}
\details{
NA values are removed from the data

Applying the function on a numerical vector should return one value

Applying the function on a factor should return nlevels + 1 value, or one value per factor level

See \code{parse_formula} for the usage for formulaes.
}
\keyword{internal}
