% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_auto}
\alias{stats_auto}
\title{Function to create a list of statistics to use in desctable}
\usage{
stats_auto(data)
}
\arguments{
\item{data}{The dataframe to apply the statistic to}
}
\value{
A list of statistics to use, assessed from the content of the dataframe
}
\description{
This function takes a dataframe as argument and returns a list of statistcs in the form accepted by desctable.
}
\details{
You can define your own automatic function, as long as it takes a dataframe as argument and returns a list of functions, or formulas defining conditions to use a stat function.
}
