% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{statTable}
\alias{statTable}
\title{Generate the table of all statistics for all variables}
\usage{
statTable(data, stats)
}
\arguments{
\item{data}{The dataframe to apply the statistic to}

\item{stats}{A list of named statistics to use}
}
\value{
A dataframe of all statistics for all variables
}
\description{
If stats is a list of functions or purrr::map like formulas, use them.
If it is a single function, use it with the entire data as
its argument to produce a list of statistical functions to use.
}
