% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{subNames}
\alias{subNames}
\title{Create the subtables names}
\usage{
subNames(grp, df)
}
\arguments{
\item{grp}{Grouping factor}

\item{df}{Dataframe containing the grouping factor}
}
\value{
A character vector with the names for the subtables
}
\description{
Create the subtables names, as
factor: level (n=sub-group length)
}
\keyword{deprecated}
\keyword{internal}
