% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{subTable}
\alias{subTable}
\title{Create a subtable in a grouped desctable}
\usage{
subTable(df, stats, tests, grps)
}
\arguments{
\item{df}{Dataframe to use}

\item{stats}{Stats list/function to use}

\item{tests}{Tests list/function to use}

\item{grps}{List of symbols for grouping factors}
}
\value{
A nested list of statTables and testColumns
}
\description{
Create a subtable in a grouped desctable
}
\keyword{deprecated}
\keyword{internal}
