% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data}
\format{A data frame with 13 rows and 7 variables}
\usage{
test_data
}
\description{
A dataset containing every type of variables encountered in statistical analyses,
with NA and NaN values as well.
}
\keyword{datasets}
