% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{testColumn}
\alias{testColumn}
\title{Create the pvalues column}
\usage{
testColumn(df, tests, grp)
}
\arguments{
\item{df}{Dataframe to use for the tests}

\item{tests}{Test function or list of functions}

\item{grp}{Grouping factor}
}
\value{
A numeric vector of pvalues
}
\description{
Create the pvalues column
}
