% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testify}
\alias{testify}
\title{Transform any test function into a valid test function for the table}
\usage{
testify(x, f, group)
}
\arguments{
\item{x}{A vector}

\item{f}{The function to try to apply, or a formula combining two functions}

\item{group}{Grouping factor}
}
\value{
The results for the function applied on the vector, compatible with the format of the result table
}
\description{
Transform a function into a valid test function for the table
Applying the function on a numerical vector should return one value
Applying the function on a factor should return nlevels + 1 value, or one value per factor level
}
