% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_auto}
\alias{tests_auto}
\title{Function to choose a statistical test}
\usage{
tests_auto(var, grp)
}
\arguments{
\item{var}{The variable to test}

\item{grp}{The variable for the groups}
}
\value{
A statistical test function
}
\description{
This function takes a variable and a grouping variable as arguments, and returns a statistcal test to use, expressed as a single-term formula.
}
\details{
This function uses appropriate non-parametric tests depending on the number of levels (wilcoxon.test for two levels
and kruskal.test for more), and fisher.test with fallback on chisq.test on error for factors.
}
