% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{varColumn}
\alias{varColumn}
\title{Generate the variable column to display as row names}
\usage{
varColumn(data, labels = NULL)
}
\arguments{
\item{data}{The dataframe to get the names from}

\item{labels}{The optional named character vector containing the keypairs var = "Label"}
}
\value{
A dataframe with one variable named "Variables", a character vector of variable names/labels and levels
}
\description{
Generates the variable column.
Replaces the variable names by their label if given in the named character vector labels, and inserts levels for factors.
}
\details{
labels is an option named character vector used to make the table prettier.
If given, the variable names for which there is a label will be replaced by their corresponding label.
Not all variables need to have a label, and labels for non-existing variables are ignored.
}
