% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.desctable}
\alias{which.desctable}
\title{Is the object possibly a desctable?}
\usage{
which.desctable(desctable)
}
\arguments{
\item{desctable}{A potential desctable to check}
}
\value{
The type of desctable or FALSE
}
\description{
Check if the object is produced by desc_table.
Return a string:
- simple
- grouped
or FALSE if not a desctable
}
\keyword{internal}
