% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{ANOVA}
\alias{ANOVA}
\title{Wrapper for oneway.test(var.equal = T)}
\usage{
ANOVA(formula)
}
\arguments{
\item{formula}{An anova formula (\code{variable ~ grouping variable})}
}
\description{
Wrapper for oneway.test(var.equal = T)
}
\seealso{
\code{\link{oneway.test}}
}
