% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headerList}
\alias{headerList}
\title{build a header list object}
\usage{
headerList(desctable)
}
\arguments{
\item{desctable}{a desctable}
}
\value{
a nested list of headers with colspans
}
\description{
build a header list object
}
