% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.normal}
\alias{is.normal}
\title{Test if distribution is normal}
\usage{
is.normal(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
A boolean
}
\description{
Test if distribution is normal.
The condition for normality is length > 30 and non-significant Shapiro-Wilks test with p > .1
}
