% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{N_Mean}
\alias{N_Mean}
\title{Return the number of observations / mean}
\usage{
N_Mean(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of means / sample sizes
}
\description{
Return the mean of a numeric vector.
NAs are removed by default.
For factors, return the sample size for each level
}

