% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Sd_Pct}
\alias{Sd_Pct}
\title{Return the percentage / sd}
\usage{
Sd_Pct(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The sd of the vector / percentages of each level of the factor
}
\description{
Return the sd of a numeric vector.
NAs are removed by default.
For factors, return the percentage of each level
}

