% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{statColumn}
\alias{statColumn}
\title{Generate one statistic for all variables}
\usage{
statColumn(stat, data)
}
\arguments{
\item{stat}{The statistic to use}

\item{data}{The dataframe to apply the statistic to}
}
\value{
A vector for one statistic column
}
\description{
Generate one statistic for all variables
}
