% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_functions.R
\name{stats_default}
\alias{stats_auto}
\alias{stats_default}
\alias{stats_nonparam}
\alias{stats_param}
\title{Functions to create a list of statistics to use in desctable}
\usage{
stats_default(data)

stats_param(data)

stats_nonparam(data)

stats_auto(data)
}
\arguments{
\item{data}{The dataframe to apply the statistic to}
}
\value{
A list of statistics to use, potentially assessed from the dataframe
}
\description{
Functions to create a list of statistics to use in desctable
}

