% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_default}
\alias{stats_default}
\alias{stats_normal}
\alias{stats_nonnormal}
\alias{stats_auto}
\title{Functions to create a list of statistics to use in desctable}
\usage{
stats_default(data)

stats_normal(data)

stats_nonnormal(data)

stats_auto(data)
}
\arguments{
\item{data}{The dataframe to apply the statistic to}
}
\value{
A list of statistics to use, potentially assessed from the dataframe
}
\description{
These functions take a dataframe as argument and return a list of statistcs in the form accepted by desctable.
}
\details{
Already defined are
- stats_default with length, mean/%, sd, med and IQR
- stats_normal with length, mean/% and sd
- stats_nonnormal with length, median/% and IQR
- stats_auto, which picks stats depending of the data

You can define your own automatic functions, as long as they take a dataframe as argument and return a list of functions or formulas defining conditions to use a stat function.
}
