% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{stats_default}
\alias{stats_default}
\alias{stats_normal}
\alias{stats_nonnormal}
\title{Define a list of default statistics}
\usage{
stats_default(data)

stats_normal(data)

stats_nonnormal(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A list of statistical functions
}
\description{
Define a list of default statistics
}
\keyword{deprecated}
