% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{statTable}
\alias{statTable}
\title{Generate the table of all statistics for all variables}
\usage{
statTable(data, stats)
}
\arguments{
\item{data}{The dataframe to apply the statistic to}

\item{stats}{A list of statistics to use}
}
\value{
A dataframe of all statistics for all variables
}
\description{
Generate the table of all statistics for all variables
}

