% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_functions.R
\name{tests_auto}
\alias{tests_auto}
\title{Functions to choose a statistical test}
\usage{
tests_auto(var, grp)
}
\arguments{
\item{var}{The variable to test}

\item{grp}{The variable for the groups}
}
\description{
Functions to choose a statistical test
}
