% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_auto}
\alias{tests_auto}
\title{Functions to choose a statistical test}
\usage{
tests_auto(var, grp)
}
\arguments{
\item{var}{The variable to test}

\item{grp}{The variable for the groups}
}
\value{
A statistical test function
}
\description{
These functions take a variable and a grouping variable as arguments, and return a statistcal test to use, expressed as a single-term formula.
}
\details{
Currently, only \code{tests_auto} is defined, and picks between t test, wilcoxon, anova, kruskal-wallis and fisher depending on the number of groups, the type of the variable, the normality and homoskedasticity of the distributions.
}
